+++ go-app/app.go
@@ -156,7 +156,7 @@ func (a *App) GetVersion() string {
 	return VERSION
 }
 
-func (a *App) GetDeviceIP() (string, error) {
+func (a *App) GetControlServerAddr() (string, error) {
 	conn, err := net.Dial("udp", "8.8.8.8:80")
 	if err != nil {
 		return "", err
@@ -164,7 +164,9 @@ func (a *App) GetDeviceIP() (string, error) {
 	defer conn.Close()
 
 	localAddr := conn.LocalAddr().(*net.UDPAddr)
-	return localAddr.IP.String(), nil
+	ip := localAddr.IP.String()
+	port := a.connector.Port()
+	return fmt.Sprintf("%s:%d", ip, port), nil
 }
 
 func (a *App) LoadConfiguration() {
@@ -203,13 +205,13 @@ func (a *App) LoadConfiguration() {
 				}
 			}
 			if calibration != nil {
-				logger.Logger.Info("[App] registering SDL map and calibration for controller", "name", sdl_mapping.Name, "usb_id", sdl_mapping.UsbID)
+				logger.Logger.Debug("[App] registering SDL map and calibration for controller", "name", sdl_mapping.Name, "usb_id", sdl_mapping.UsbID)
 				a.sdl_controller_manager.RegisterConfig(sdl_mapping, *calibration)
 			}
 		}
 
 		for _, profile := range profiles {
-			logger.Logger.Info("[App] registering profile", "profile", profile.Id(), profile.Name)
+			logger.Logger.Debug("[App] registering profile", "profile", profile.Id(), "name", profile.Name)
 			a.profile_runner.RegisterProfile(profile)
 		}
 	}