+++ go-app/config_loader/config_loader.go
@@ -33,9 +33,6 @@ func (c *ConfigLoader) FromFS(fsys fs.FS, options ConfigLoader_FromFS_Options) (
 	sdl_mapping_files_dir := path.Join(DIR_SDL_MAPPINGS_NAME)
 	profiles_files_dir := path.Join(DIR_PROFILES_NAME)
 
-	entries, _ := fs.ReadDir(fsys, "")
-	fmt.Printf("%#v\n", entries)
-
 	calibration_file_entries, err := fs.ReadDir(fsys, calibration_files_dir)
 	var parsed_calibration_files []config.Config_Controller_Calibration
 	if err != nil {