+++ go-app/frontend/src/swr/index.ts
@@ -6,5 +6,5 @@ export * from "./useCabControlState";
 export * from "./useSharedProfiles";
 export * from "./useLastInstalledModVersion";
 export * from "./useProfiles";
-export * from "./useDeviceIP";
+export * from "./useControlServerAddr";
 export * from "./useControllerConfiguration";