+++ go-app/tswconnector/shared.go
@@ -17,6 +17,7 @@ type TSWConnector interface {
 	Start() error
 	Stop() error
 	IsActive() bool
+	Port() int
 	Subscribe() (chan TSWConnector_Message, func())
 	Send(m TSWConnector_Message) error
 }